spatch does not always do the right thing with -dir directive
It seems to miss .h files sometimes.

I hope this is the last iteration of fixing types.

+++ sys/include/ape/signal.h
@@ -64,7 +64,7 @@ extern int raise(int);
 
 #ifdef _POSIX_SOURCE
 
-typedef long sigset_t;
+typedef int32_t sigset_t;
 struct sigaction {
 	void		(*sa_handler)();
 	sigset_t	sa_mask;
